feat(cart): 장바구니 추가 기능 백엔드 API 구현
- JWT 인증 필터(`JwtRequestFilter`)를 추가하여 API 엔드포인트 보안 적용
- `PrincipalDetails`를 사용해 인증된 사용자 정보를 컨트롤러에 주입
- 장바구니 및 장바구니 아이템 추가/수정 로직을 `CartService`에 구현
- 상품/회원 조회 실패 시 사용할 `ResourceNotFoundException` 추가

+++ academy_project_backend/src/main/java/com/wnsdudwh/Academy_Project/entity/Member.java
@@ -19,7 +19,7 @@ public class Member
 {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idx;
+    private Long idx;
 
     @Column(nullable = false)
     private String usernumber;
